+++ main-service/src/main/java/ru/practicum/service/PublicService.java
@@ -6,13 +6,13 @@ import ru.practicum.dto.event.EventsPublicSearchDto;
 public interface PublicService {
     ResponseEntity<Object> getCompilations(Boolean pinned, Integer from, Integer size);
 
-    ResponseEntity<Object> getCompilationById(Integer compId);
+    ResponseEntity<Object> getCompilationById(Long compId);
 
     ResponseEntity<Object> getCategories(Integer from, Integer size);
 
-    ResponseEntity<Object> getCategoryById(Integer catId);
+    ResponseEntity<Object> getCategoryById(Long catId);
 
     ResponseEntity<Object> getEvents(EventsPublicSearchDto searchDto);
 
-    ResponseEntity<Object> getEventById(Integer id);
+    ResponseEntity<Object> getEventById(Long id);
 }
